pyproject.toml
[tool.poetry]
name = "BrokRPC"
version = "0.3.0"
description = "framework for gRPC like server-client communication over message brokers"
authors = ["zerlok <danil.troshnev@gmail.com>"]
readme = "README.md"
license = "MIT"
keywords = [
    "python",
    "protobuf",
    "amqp",
    "grpc",
    "message-queue",
    "message-broker",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: System :: Networking",
    "Typing :: Typed",
]

[tool.poetry.urls]
Homepage = "https://github.com/zerlok/BrokRPC"
Issues = "https://github.com/zerlok/BrokRPC/issues"

[tool.poetry.dependencies]
python = "^3.12"
yarl = "^1.18.0"
protobuf = {version = "^5.26.1", optional = true}
googleapis-common-protos = {version = "^1.65.0", optional = true}
aiormq = {version = "^6.8.1", optional = true}
aiofiles = {version = "^24.1.0", optional = true}
pydantic = {version = "^2.10.4", optional = true}

[tool.poetry.extras]
cli = ["aiofiles"]
protobuf = ["protobuf", "googleapis-common-protos"]
aiormq = ["aiormq"]
pydantic = ["pydantic"]

[tool.poetry.scripts]
brokrpc = "brokrpc.cli:main"

[tool.poetry.group.dev.dependencies]
types-protobuf = "^5.26.0.20240422"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-cov = ">=5,<7"
ruff = "^0.8.0"
pytest-asyncio = ">=0.24,<0.26"
pytest-timeout = "^2.3.1"
pytest-mypy-plugins = "^3.1.2"
types-aiofiles = "^24.1.0.20240626"
pyprotostuben = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
include = ["src/**/*.py", "tests/**/*.py"]
exclude = ["**_pb2.py", "**_pb2.pyi", "**_pb2_grpc.py", "**_pb2_grpc.pyi"]
line-length = 120
output-format = "pylint"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN", # because we use mypy
    "D", # TODO: add docstrings to public code
    "FA", # TODO: consider should we use __annotations__
    "TD", # no task tracking
    "FIX", # TODO: consider enable it against new code on pull requests
    "A005", # TODO: wait for ruff 0.10.0
    "COM812", # because ruff format suggests to skip it
    "ISC001", # because ruff format suggests to skip it
    "RET505", # clashes with mypy exhaustiveness check
    "S101", # allow asserts for tests checks and mypy help
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "ARG001", # it's ok to use a fixture with a side effect in a test.
    "PT004", # fixture may add side effect and doesn't return anything.
]


[tool.mypy]
files = ["src", "tests", "examples/simple.py", "examples/simple_rpc.py"]

strict = true
disallow_any_unimported = true
disallow_any_expr = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true
strict_equality = true
strict_optional = true
enable_error_code = [
    "redundant-self",
    "redundant-expr",
    "possibly-undefined",
    "truthy-bool",
    "truthy-iterable",
    "ignore-without-code",
    "unused-awaitable",
    # TODO: enable explicit override
    #  "explicit-override",
    "mutable-override",
    "unimported-reveal",
    "narrowed-type-not-subtype",
]

# NOTE: google.rpc doesn't have appropriate types
[[tool.mypy.overrides]]
module = ["google.rpc.*"]
ignore_missing_imports = true

# NOTE: google.rpc doesn't have appropriate types, but these types are used by call_pb2
[[tool.mypy.overrides]]
module= ["brokrpc.spec.v1.call_pb2"]
disallow_any_unimported = false

# NOTE: in the following modules it is hard to avoid expressions with Any
[[tool.mypy.overrides]]
module= [
    "brokrpc.builder",
    "brokrpc.cli",
    "brokrpc.debug",
    "brokrpc.driver.aiormq",
    "brokrpc.options",
    "brokrpc.plugin",
    "brokrpc.rpc.server",
]
disallow_any_expr = false

# NOTE: allow return `typing.Any` in test fixtures (e.g. mock objects created with `create_autospec`)
[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_any_expr = false
disallow_any_explicit = false
warn_return_any = false

# NOTE: pydantic.BaseModel has fields with Any
[[tool.mypy.overrides]]
module= ["tests.stub.pydantic"]
disallow_any_decorated = false


[tool.pytest.ini_options]
pythonpath = [
    "src",
]
addopts = [
    "--timeout=60", # 1 minute for each test
    "--cov=src",
    "--cov-report=term-missing",
]
testpaths = [
    "tests",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_also = [
    "if __name__ == .__main__.:",
    "if t\\.TYPE_CHECKING:",
    "class .*\\(t\\.Protocol\\):",
    "@abc\\.abstractmethod",
    "t\\.assert_never\\(.*\\)",
]
show_missing = true
